requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
brotli==1.2.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.3
fsspec==2026.1.0
gradio==6.5.1
gradio-client==2.0.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.3.5
idna==3.11
jinja2==3.1.6
jiter==0.12.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
numpy==2.4.1
openai==2.16.0
orjson==3.11.6
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
pyyaml==6.0.3
rich==14.3.1
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uvicorn==0.40.0
